backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# CORS and Filtering
django-cors-headers==4.3.1
django-filter==23.5

# Image Processing
Pillow==10.2.0

# HTTP Requests (for Naver Map API)
requests==2.31.0

# Environment Variables
python-decouple==3.8

# Development/Testing
faker==22.0.0
